scaling number for offers
This notebook contains simple methods of scaling metrics for the offers dashbaord so they are closer to the metrics we would see across the entire retailer. There are a lot of limitations and assumptions to doing this (mainly driven from the fact we don't have any sort of customer identifier in unmatched transactions) so we have to be careful to assume any metric thatre requires customer volume are estimates. Transacitons and redemptions can be taken directly from unmatched transactions.